Enhance order processing with validation for sell transactions and improve UI interaction for order submission

+++ routes/trade_routes.py
@@ -1,3 +1,4 @@
+from decimal import Decimal as decimal
 from flask import Blueprint, jsonify, request,session
 from service.trade_service import place_order
 
@@ -14,7 +15,7 @@ def place_trade_order():
         quantity = int(quantity) if quantity else 0
         order_type = data.get("order_type")
         price = data.get("price")  
-        price = float(price) if price else None
+        price = decimal(price) if price else None
         transaction_type = data.get("transaction_type")
 
         result = place_order(
